pyproject.toml
[project]
name = "agario-rl-experiment"
version = "0.1.0"
description = "Minimal multi-agent Agar.io-style RL lab with supervisor controls."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
  "numpy>=1.26",
  "raylib>=5.5.0.4",
  "torch>=2.2",
  "PyYAML>=6.0",
  "matplotlib>=3.8",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
license-files = ["LICENSE"]

[tool.setuptools.packages.find]
include = ["agario_rl*"]
